[codecache][triton 3.2] hash -> base64 conversion for triton 3.2
In old triton versions, you take the hash of the triton kernel and use it in the filepath for the cached kernel. In Triton 3.2 (after https://github.com/triton-lang/triton/pull/4553), the filepath will use the base-64-encoded representation of the hash in the path.

This PR checks whether the `_base64` function exists in triton, and if so, uses the base-64-encoded represenatation in the path.

ghstack-source-id: f0e880d5ac1b539841d19d126a8e194760382c87
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140190