[Inductor] Expand dtype aware codegen for unary ops (#139839)
Summary:

Previously, only `torch.sqrt` was dtype aware. This PR updates most of the unary ops to have it as well. This is necessary to get correct code for libdevice and tl.math ops, as their inputs need to be upcasted to float32.

Test Plan: Added CI tests for all the new ops.

Differential Revision: D65517197